upgrade .net 2, 4.7, readme, start with ranking towards https://string-db.org/cgi/input.pl?input_page_active_form=proteins_with_values

+++ PerseusPluginLib/Impute/ReplaceImputedByNan.cs
@@ -39,7 +39,7 @@ namespace PerseusPluginLib.Impute{
 			for (int i = 0; i < data.RowCount; i++){
 				for (int j = 0; j < data.ColumnCount; j++){
 					if (data.IsImputed[i, j]){
-						data.Values.Set(i, j, float.NaN);
+						data.Values.Set(i, j, double.NaN);
 					}
 				}
 			}
